+++ package.json
@@ -30,6 +30,7 @@
     "multer": "^1.4.2",
     "nodemailer": "^6.4.2",
     "passport": "^0.4.1",
+    "passport-google-oauth": "^2.0.0",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "pino": "^5.16.0",
